+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- TBD
+### Added
+
+- `get_source_tree` function in audit.py
 
 
 ## [1.5.0] - 2025-10-10